Complete JSON creation
