+++ Cargo.toml
@@ -9,6 +9,7 @@ edition = "2021"
 aead = "0.5.2"
 aes = "0.8.3"
 aes-gcm = { version = "0.10.2", features = ["heapless"] }
+async-std = "1.12.0"
 base64 = "0.21.2"
 bip39 = "2.0.0"
 block-modes = "0.8"
