+++ src/stored_procedure/mod.rs
@@ -1,4 +1,2 @@
-pub mod accountquery;
-pub mod blindassetrecord;
 pub mod keys;
 pub mod record;
